move display components to central location
