+++ apps/web/src/components/WorkforceAdjustment/Warning.tsx
