+++ apps/web/src/messages/workforceAdjustmentMessages.ts
@@ -1,0 +1,29 @@
+import { defineMessages } from "react-intl";
+
+export default defineMessages({
+  wfa: {
+    defaultMessage: "Workforce adjustment",
+    id: "xGsQso",
+    description: "Title for the users wfa information",
+  },
+  wfaSituation: {
+    defaultMessage: "Workforce adjustment situation",
+    id: "W9/kLq",
+    description: "Label for users current wfa situation",
+  },
+  currentSubstantivePosition: {
+    defaultMessage: "Current substantive position",
+    id: "3pK4nh",
+    description: "Label for users current substantive position",
+  },
+  expectedEndDate: {
+    defaultMessage: "Expected end date",
+    id: "JSyD03",
+    description: "Label for the expected end date",
+  },
+  currentCommunity: {
+    defaultMessage: "Current functional community",
+    id: "OU3MYy",
+    description: "Label for users current functional community",
+  },
+});
