+++ apps/web/src/pages/EmployeeProfile/messages.ts
@@ -1,6 +1,11 @@
 import { defineMessages } from "react-intl";
 
 export default defineMessages({
+  wfa: {
+    defaultMessage: "Workforce adjustment",
+    id: "xGsQso",
+    description: "Title for the users wfa information",
+  },
   careerDevelopment: {
     defaultMessage: "Career development preferences",
     id: "okKbgd",
@@ -26,30 +31,5 @@ export default defineMessages({
     id: "DX3z2x",
     description:
       "Selection or label referring to adding a community that isn't a provided option",
-  },
-  wfa: {
-    defaultMessage: "Workforce adjustment",
-    id: "xGsQso",
-    description: "Title for the users wfa information",
-  },
-  wfaSituation: {
-    defaultMessage: "Workforce adjustment situation",
-    id: "W9/kLq",
-    description: "Label for users current wfa situation",
-  },
-  currentSubstantivePosition: {
-    defaultMessage: "Current substantive position",
-    id: "3pK4nh",
-    description: "Label for users current substantive position",
-  },
-  expectedEndDate: {
-    defaultMessage: "Expected end date",
-    id: "JSyD03",
-    description: "Label for the expected end date",
-  },
-  currentCommunity: {
-    defaultMessage: "Current functional community",
-    id: "OU3MYy",
-    description: "Label for users current functional community",
   },
 });
